requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
google-cloud==0.34.0
google-cloud-bigquery==3.13.0
google-api-python-client==2.108.0
db-dtypes==1.1.1